refactor: isolate runs reconciliation
